// server/src/enums/business.enum.ts
export enum BusinessProfileFilterField {
    BusinessCategoryUuid = 'businessCategoryUuid',
    Country = 'country',
    StateAndProvince = 'stateAndProvince',
    City = 'city'
}

export enum SortBy {
    CreatedUtc = 'createdUtc',
    ModifiedUtc = 'modifiedUtc'
}

export enum SortDirection {
    ASC = 'asc',
    DESC = 'desc'
}